major redesign

+++ webserver/USB_serial.py
@@ -1,7 +1,7 @@
 """
 name : serial.py
 Author: DK1RI
-Version 01.0, 20230406
+Version 01.0, 20240107
 call with: serial.py <comport>
 Purpose :
 interface between webserver and USB devices
@@ -36,23 +36,28 @@ ser = serial.Serial(comport, 19200, timeout=0.5)
 to_web = dir + "/to_web"
 from_web = dir + "/from_web"
 while 1:
+    error = 0
     if os.path.exists(from_web):
         f = open(from_web)
         dat = f.readline()
         f.close()
         if len(dat) != 0:
             print ("received: " + dat)
-            ser_write(ser, bytes.fromhex(dat))
-
-            in_dat = ser_read(ser)
-            if len(in_dat) != 0:
-                if os.path.exists(to_web):
-                    os.remove(to_web)
-                f = open(to_web, "a")
-                send = in_dat.hex()
-                f.write(send)
-                f.close()
-                print ("sent: ")
-                print(send)
+            try:
+                ser_write(ser, bytes.fromhex(dat))
+            except:
+                print("hex error")
+                error = 1
+            if error == 0:
+                in_dat = ser_read(ser)
+                if len(in_dat) != 0:
+                    if os.path.exists(to_web):
+                        os.remove(to_web)
+                    f = open(to_web, "a")
+                    send = in_dat.hex()
+                    f.write(send)
+                    f.close()
+                    print ("sent: ")
+                    print(send)
         os.remove(from_web)
     time.sleep(1)
